🗃 Enhance submissions system to show all content at once
