+++ hasura/metadata/actions.yaml
@@ -29,6 +29,17 @@ actions:
       value_from_env: EVENT_SECRET
   permissions:
   - role: user
+- name: getProgramPersonAccessToken
+  definition:
+    kind: ""
+    handler: '{{ACTION_BASE_URL}}/element/getProgramPersonAccessToken'
+    forward_client_headers: true
+    headers:
+    - name: x-hasura-event-secret
+      value_from_env: EVENT_SECRET
+  permissions:
+  - role: user
+  - role: unauthenticated
 - name: getSlug
   definition:
     kind: ""
@@ -245,4 +256,5 @@ custom_types:
   - name: SubmitUploadableElementOutput
   - name: NotifyEventEnded
   - name: GetSlugOutput
+  - name: MatchingPersonOutput
   scalars: []
